requirements.txt
streamlit==1.39.0
pandas>=2.1
PyYAML>=6.0
